Moved redirect and error handling logic into `getSafeNetworkResponse`
