--- app/assets/js/src/service-worker/error/index.ts
@@ -1,1 +1,0 @@
-export { getFallbackResponse } from './getFallbackResponse';
