+++ app/assets/js/src/service-worker/strategy/networkFirst.ts
@@ -1,13 +1,11 @@
 import { addToCache, getCachedResponse } from '../cache';
-import { getFallbackResponse } from '../error';
+import { getFallbackResponse, getSafeNetworkResponse } from '../utils';
 
 /**
  * Attempt to return a regular network `Response`, but if something goes
  * wrong return a cached `Response` instead, if possible.
- *
- * The returned promise will resolve to `null` if the fetched response is an opaque redirect.
  */
-export async function networkFirst({ preloadResponse, request }: FetchEvent): Promise<Response | null> {
+export async function networkFirst({ preloadResponse, request }: FetchEvent): Promise<Response> {
 	try {
 		const networkResponse = await (async () => {
 			// If a response was preloaded, return it
@@ -17,13 +15,8 @@ export async function networkFirst({ preloadResponse, request }: FetchEvent): Pr
 			}
 
 			// Otherwise, make the request and return its response
-			return await fetch(request);
+			return await getSafeNetworkResponse(request);
 		})();
-
-		// If the response is an opaque redirect, let the browser handle it
-		if (networkResponse.type === 'opaqueredirect') {
-			return null;
-		}
 
 		// If the response arrived, cache it if it was okay, and return
 		// it immediately without waiting for it to be added to the cache
