+++ app/assets/js/src/service-worker/utils/getFallbackResponse.ts
