+++ app/assets/js/src/service-worker/utils/index.ts
@@ -1,0 +1,2 @@
+export { getFallbackResponse } from './getFallbackResponse';
+export { getSafeNetworkResponse } from './getSafeNetworkResponse';
